+++ src/deprecated/index.ts
@@ -1,7 +1,0 @@
-export { Vec2 } from "./vec2";
-export { SmallIntCache } from "./small-int-cache";
-export { Map1 } from "./map1";
-export { Map2 } from "./map2";
-export { Map3 } from "./map3";
-export { SetBase, Set1, DeepSet } from "./set";
-export { DivRect } from "./div-rect";