+++ src/utils/index.ts
@@ -5,11 +5,3 @@ export * as Map from "./map";
 export * as Math from "./math";
 export * as Obj from "./obj";
 export * as Str from "./str";
-
-import * as Guard from "../guard";
-
-/**
- * @deprecated Use {@link Guard} instead. Will be removed in v2.0.0.
- * @private
- */
-export const Is = Guard;