+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## [2.0.0] - 2025-10-31
+### Breaking Major Update
+- Removed all deprecated functions and classes.
+- Assert update, so that assertions can be used `let i =  Assert.isInteger(i);`.
 
 ## [1.21.0] - 2025-10-30
 ### Added